Fixed the skripts everything downloaded and updated path to the raw data .gitignore

+++ src/data/dataGrabber.py
@@ -1,46 +1,52 @@
 import pandas as pd
 import requests
-import time
+import os.path
 
 #helperfunctions
 #----------------------
+#https://www.basketball-reference.com/teams/ATL/1992/gamelog/#sum:tgl_basic
 
 #gets all tables from basketball-reference.com for a certain year and team 
 def getTeamYear(team, year):
-    print("debug12")
     print(team,year)
 
     link = 'https://www.basketball-reference.com/teams/' + str(team) + '/' + str(year) + '/gamelog/#sum:tgl_basic'
+    print(link)
     request = requests.get(link)
-    filename = 'C:\cygwin64\home\Tim\pythonstuff\cookiecutter-data-science\Predict_NBA\data\raw' + str(team) +str(year) + '.csv'
-    if request.status_code == 200:
-        sauce = pd.read_html(link, header = 0)
-        for df in sauce:
-            df.to_csv(filename , encoding='utf-8', index=False)
-            return;
-    else:
-        print('Web page error')
-        print(link)
-    return;
+    filename = 'C:\\cygwin64\\home\\Tim\\pythonstuff\\cookiecutter-data-science\\Predict_NBA\\data\\raw\\' + str(team) +str(year) + '.csv'
+    
+    if os.path.isfile(filename) == 0:    
+        if request.status_code == 200:
+            try: 
+                sauce = pd.read_html(link, header = 0)
+                for df in sauce:
+                    df.to_csv(filename, encoding='utf-8', index=False)
+                    return;
+            
+            except ValueError:
+                print("No tables at:",link)
+        else:
+            print('Web page error')
+            print(link)
+        return;
 
 #getTeamnames that are relevant during that year
 def getAll():
-    teams = pd.read_csv('teamNamesPerYear.csv',index_col=0)
-    print("debug1")
+    teams = pd.read_csv('teamNamesPerYear.csv',index_col=0).T
     
-    for index, row in teams.iterrows():
-            getTeamYear(str(index),str(row))
-            #print(row, index)
-            time.sleep(1)
-    
+    for row in teams.iterrows():
+        year = str(row[0])
+        series = row[1]
+        for team in series:
+            if pd.isnull(team) == 0 :   
+                getTeamYear(team,year)
     return;
     
 
     
     
-#script
-teams = pd.read_csv('teamNamesPerYear.csv',index_col=0)
+#script 
+getAll()
 
-#getTeamYear(ATL,1980)
 
-C:\cygwin64\home\Tim\pythonstuff\cookiecutter-data-science\Predict_NBA\data\raw
+
